move iceberg sql into a separate module
