+++ sdks/java/extensions/sql/build.gradle
@@ -75,12 +75,6 @@ dependencies {
   fmppTemplates library.java.vendored_calcite_1_40_0
   implementation project(path: ":sdks:java:core", configuration: "shadow")
   implementation project(":sdks:java:managed")
-  if (JavaVersion.current().compareTo(JavaVersion.VERSION_11) >= 0) {
-    // iceberg ended support for Java 8 in 1.7.0
-    implementation project(":sdks:java:io:iceberg")
-    runtimeOnly project(":sdks:java:io:iceberg:bqms")
-    runtimeOnly project(":sdks:java:io:iceberg:hive")
-  }
   implementation project(":sdks:java:extensions:avro")
   implementation project(":sdks:java:extensions:join-library")
   permitUnusedDeclared project(":sdks:java:extensions:join-library") // BEAM-11761
@@ -121,8 +115,6 @@ dependencies {
   permitUnusedDeclared library.java.hadoop_client
   provided library.java.kafka_clients
 
-  testImplementation "org.apache.iceberg:iceberg-api:1.6.1"
-  testImplementation "org.apache.iceberg:iceberg-core:1.6.1"
   testImplementation library.java.vendored_calcite_1_40_0
   testImplementation library.java.vendored_guava_32_1_2_jre
   testImplementation library.java.junit
