+++ sdks/java/extensions/sql/iceberg/build.gradle
@@ -1,0 +1,44 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * License); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an AS IS BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+plugins { id 'org.apache.beam.module' }
+
+applyJavaNature(
+  automaticModuleName: 'org.apache.beam.sdk.extensions.sql.meta.provider.hcatalog',
+  // iceberg ended support for Java 8 in 1.7.0
+  requireJavaVersion: JavaVersion.VERSION_11,
+)
+
+dependencies {
+  implementation project(":sdks:java:extensions:sql")
+  implementation project(":sdks:java:core")
+  implementation project(":sdks:java:io:iceberg")
+  runtimeOnly project(":sdks:java:io:iceberg:bqms")
+  runtimeOnly project(":sdks:java:io:iceberg:hive")
+  implementation library.java.slf4j_api
+  implementation library.java.vendored_guava_32_1_2_jre
+
+  testImplementation library.java.joda_time
+  testImplementation library.java.junit
+  testImplementation library.java.vendored_calcite_1_40_0
+  testImplementation library.java.google_api_services_bigquery
+  testImplementation "org.apache.iceberg:iceberg-api:1.9.2"
+  testImplementation "org.apache.iceberg:iceberg-core:1.9.2"
+  testImplementation project(":sdks:java:io:google-cloud-platform")
+  testImplementation project(":sdks:java:extensions:google-cloud-platform-core")
+}
