+++ settings.gradle.kts
@@ -372,3 +372,5 @@ include("sdks:java:io:iceberg:bqms")
 findProject(":sdks:java:io:iceberg:bqms")?.name = "bqms"
 include("it:clickhouse")
 findProject(":it:clickhouse")?.name = "clickhouse"
+include("sdks:java:extensions:sql:iceberg")
+findProject(":sdks:java:extensions:sql:iceberg")?.name = "iceberg"
